backend/requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
Django==5.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
Flask==3.0.3
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qrcode==8.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.0.4
